chore: create separate singleUser subreducer
